[ISSUE-515] Rename node operator image
