+++ Makefile.docker
@@ -9,7 +9,7 @@ image-controller \
 image-extender \
 image-extender-patcher \
 image-scheduler \
-image-operator
+image-node-controller
 
 base-images: base-image-drivemgr base-image-node base-image-controller
 
@@ -61,11 +61,10 @@ image-scheduler:
 	cp ./pkg/${SCHEDULER}/${PLUGIN}/Dockerfile ./build/${SCHEDULING_PKG}/${SCHEDULER}/
 	docker build --network host --force-rm --tag ${REGISTRY}/${PROJECT}-${SCHEDULER}:${TAG} ./build/${SCHEDULING_PKG}/${SCHEDULER}
 
-image-operator:
-	cp -r ./${DRIVER_CHART_PATH} ./build/${CR_CONTROLLERS}/${OPERATOR}/
-	cp ./pkg/${CR_CONTROLLERS}/${OPERATOR}/Dockerfile ./build/${CR_CONTROLLERS}/${OPERATOR}/
-	docker build --network host --force-rm --tag ${REGISTRY}/${PROJECT}-${OPERATOR}:${TAG} \
-	./build/${CR_CONTROLLERS}/${OPERATOR}
+image-node-controller:
+	cp ./pkg/${CR_CONTROLLERS}/${NODE_CONTROLLER_PKG}/Dockerfile ./build/${CR_CONTROLLERS}/${NODE_CONTROLLER}/
+	docker build --network host --force-rm --tag ${REGISTRY}/${PROJECT}-${NODE_CONTROLLER}:${TAG} \
+	./build/${CR_CONTROLLERS}/${NODE_CONTROLLER}
 
 ###################
 ### Push images ###
@@ -76,7 +75,7 @@ push-controller \
 push-extender \
 push-extender-patcher \
 push-scheduler \
-push-operator
+push-node-controller
 
 push-drivemgr:
 	docker push ${REGISTRY}/${PROJECT}-${DRIVE_MANAGER_TYPE}:${TAG}
@@ -97,8 +96,8 @@ push-extender-patcher:
 push-scheduler:
 	docker push ${REGISTRY}/${PROJECT}-${SCHEDULER}:${TAG}
 
-push-operator:
-	docker push ${REGISTRY}/${PROJECT}-${OPERATOR}:${TAG}
+push-node-controller:
+	docker push ${REGISTRY}/${PROJECT}-${NODE_CONTROLLER}:${TAG}
 
 ####################
 ### Clean images ###
@@ -109,7 +108,7 @@ clean-image-drivemgr \
 clean-image-extender \
 clean-image-extender-patcher \
 clean-image-scheduler \
-clean-image-operator
+clean-image-node-controller
 
 clean-image-drivemgr:
 	docker rmi ${REGISTRY}/${PROJECT}-${DRIVE_MANAGER_TYPE}:${TAG}
@@ -130,5 +129,5 @@ clean-image-extender-patcher:
 clean-image-scheduler:
 	docker rmi ${REGISTRY}/${PROJECT}-${SCHEDULER}:${TAG}
 
-clean-image-operator:
-	docker rmi ${REGISTRY}/${PROJECT}-${OPERATOR}:${TAG}
+clean-image-node-controller:
+	docker rmi ${REGISTRY}/${PROJECT}-${NODE_CONTROLLER}:${TAG}
